Finished the delete method to a working standard. Altered the aspx pages for easier use.
