+++ FolioFront/FolioFront.csproj
@@ -68,10 +68,10 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="about.aspx" />
+    <Content Include="addEditDelete\AddAuthor.aspx" />
     <Content Include="addEditDelete\AuthorAddEditDelete.aspx" />
     <Content Include="addEditDelete\AuthorDeleteConfirm.aspx" />
     <Content Include="assets\css\primer.css" />
-    <Content Include="AuthorList.aspx" />
     <Content Include="find\author.aspx" />
     <Content Include="find\customer.aspx" />
     <Content Include="default.aspx" />
@@ -91,6 +91,13 @@
     <Compile Include="about.aspx.designer.cs">
       <DependentUpon>about.aspx</DependentUpon>
     </Compile>
+    <Compile Include="addEditDelete\AddAuthor.aspx.cs">
+      <DependentUpon>AddAuthor.aspx</DependentUpon>
+      <SubType>ASPXCodeBehind</SubType>
+    </Compile>
+    <Compile Include="addEditDelete\AddAuthor.aspx.designer.cs">
+      <DependentUpon>AddAuthor.aspx</DependentUpon>
+    </Compile>
     <Compile Include="addEditDelete\AuthorAddEditDelete.aspx.cs">
       <DependentUpon>AuthorAddEditDelete.aspx</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
@@ -104,13 +111,6 @@
     </Compile>
     <Compile Include="addEditDelete\AuthorDeleteConfirm.aspx.designer.cs">
       <DependentUpon>AuthorDeleteConfirm.aspx</DependentUpon>
-    </Compile>
-    <Compile Include="AuthorList.aspx.cs">
-      <DependentUpon>AuthorList.aspx</DependentUpon>
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
-    <Compile Include="AuthorList.aspx.designer.cs">
-      <DependentUpon>AuthorList.aspx</DependentUpon>
     </Compile>
     <Compile Include="default_master.Master.cs">
       <DependentUpon>default_master.Master</DependentUpon>
