+++ FolioFront/addEditDelete/AuthorAddEditDelete.aspx.cs
@@ -13,44 +13,28 @@ namespace FolioFront
         Int32 AuthorId;
         protected void Page_Load(object sender, EventArgs e)
         {
-
-        }
-        void Add()
-        {
-            ClsAuthorCollection AuthorCollection = new ClsAuthorCollection();
-            string authorName = txtAEDAuthorName.Text;
-            string authorDob = txtAEDAuthorDob.Text;
-            String Error = AuthorCollection.ThisAuthor.AuthorNameValid(authorName) + ". " +
-                           AuthorCollection.ThisAuthor.AuthorDobValid(authorDob);
-            if(Error == "")
-            { 
-                AuthorCollection.ThisAuthor.Name = txtAEDAuthorName.Text;
-                AuthorCollection.ThisAuthor.DOB = Convert.ToDateTime(authorDob);
-                AuthorCollection.ThisAuthor.IsAlive = Convert.ToBoolean(cbxAEDAuthorIsAlive.Checked);
-                AuthorCollection.Add();
-                Response.Redirect("AuthorAddEditDelete.aspx");
-            }
-            else
+            if (IsPostBack == false)
             {
-                lblErrorAED.Text = "There were issues with data entered";                
+                DisplayAuthors();
             }
-
         }
-        void DeleteAuthor()
+        void DisplayAuthors()
         {
-            ClsAuthorCollection AllAuthors = new ClsAuthorCollection();
-            AllAuthors.ThisAuthor.Find(AuthorId);
-            AllAuthors.Delete();
+            FolioClasses.ClsAuthorCollection Authors = new FolioClasses.ClsAuthorCollection();
+            lstAuthors.DataSource = Authors.AuthorList;
+            lstAuthors.DataValueField = "AuthorId";
+            lstAuthors.DataTextField = "Name";
+            lstAuthors.DataBind();
         }
+        
         protected void btnAddAuthor_Click(object sender, EventArgs e)
         {
-            Add();
-            Response.Redirect("AuthorAddEditDelete.aspx");
+            Session["AuthorId"] = -1;
+            Response.Redirect("AddAuthor.aspx");
         }
         protected void btnDeleteAuthor_Click(object sender, EventArgs e)
         {
-            Int32 AuthorId;
-            if(lstAuthors.SelectedIndex != -1)
+            if (lstAuthors.SelectedIndex != -1)
             {
                 AuthorId = Convert.ToInt32(lstAuthors.SelectedValue);
                 Session["AuthorId"] = AuthorId;
@@ -61,10 +45,6 @@ namespace FolioFront
                 lblErrorAED.Text = "Please select a record to delete from the list";
             }
         }
-        protected void btnDeleteYes_Click(object sender, EventArgs e)
-        {
-            DeleteAuthor();
-            Response.Redirect("AuthorAddEditDelete.aspx");
-        }
+        
     }
 }
